requirements.txt
altgraph==0.17.4
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
importlib_metadata==8.5.0
macholib==1.16.3
packaging==24.0
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
PyYAML==6.0.2
requests==2.32.3
scapy==2.5.0
urllib3==2.2.1
zipp==3.21.0
